CXX-Interop: Fix missing header deps edge
The custom target dependency was an order-only dependency edge. As a
result, changing the swift file wouldn't result in the header getting
regenerated at the right time because nothing actually depended on it
being up-to-date. This patch should fix that and ties it into the
`fibonacci` target a little more cleanly. The generated header is marked
as being one of the target sources, so the target itself depends on the
header getting built. This ensures that the C++ file gets rebuilt if the
header changes. The header still depends on the Swift sources in the
target. We're now extracting that list directly from the target itself
instead of needing to pass it in explicitly. The module is also
extracted from the target itself. We're also automatically adding the
location of the header include directory so that dependees can find it
as part of the target interface.

Fixes #8